Implement the virtual table data structure for classes
So for function conversion, I need to know offsets using the virtual
tables, so I had to add a new HashMap to record their offsets for
each class. It would have been prettier to implement if offset
calculation was done at runtime, but I thought the tradeoff to have
the virtual tables all calculated at the beginning to be worth it.