Refactored class parse code
Added in the functions startParseClass and endParseClass so that I could
maintain the entry and exit class code in one centralized place. I added
this to the three class nodes. Also, I included a new variable
"variable_map" in the environment, my solution for keeping track of
variables in the environment.