Change handling of a method being called
One of the bugs that got me last time was having a function call within
a function call. This is becaue I would save parameters directly into a
variable (because I was too lazy to change the NodeList visitor), and it
would get overwritten. I used a stack to fix this - push whenever we're
starting a new function call, and pop whenever we're exiting the
function call.