Implement null ptr and fixed array checking
Originally array checking was done in a way such that an access to the
exact size of the array would be allowed (so my error checking was off
by one). I had to change my code to use separate if0 and if statement
generators, and used those to do checking. Also, I added the nullptr
checking for calling methods of a class.